Add upstream "chat bot" code
Copies the upstream chat bot code from upstream. This code was removed in [[https://secure.phabricator.com/D17756 | D17756: Remove all "Phabricator Bot" code]]. The reason that it was removed upstream is:

> We aren't maintaining this at all (see, e.g., [[https://secure.phabricator.com/T7829 | T7829]]) and a user reported a technically accurate security issue via [[https://hackerone.com/reports/222870 | HackerOne]]).

Basically, the chat bot doesn't seem to respect policies and, as such, it is possible for a user that should not be able to view an object (such as a Maniphest task) to learn details about that object by querying the chat bot. Given that we don't often restrict object policies, I'm not overly concerned by this issue.