Remove "command" from chatbot messages
Summary: Ref T19770. Ref T34787. After D65641, the `$message` parameter of the `PhabricatorChatbotMessage` class is essentially useless because it is always set to `'MESSAGE'`. To be honest, I've never really understood the purpose of this parameter anyway... I think that if we want to associate different meanings with `PhabricatorChatbotMessage`s then it would be better to use subclasses rather than a class parameter.

Test Plan: `grep` for `getCommand` and `setCommand`.

Reviewers: #phabricator, jbeck, tpham

Reviewed By: tpham

Maniphest Tasks: T34787, T19770

Differential Revision: https://phabricator.freelancer.com/D65775