Add a Herald field to match user and repository projects
Summary: Ref T21277. This adds a Herald field to check if a user's projects intersect with a repositories projects. This would be used in a Herald rule to assign push permissions to repositories.

Test Plan: Put on the staging repository hosts, configured: https://phabricator.tools-stg.flnltd.com/H7, and pushed both with and without a common project with the repository.

Reviewers: josh

Reviewed By: josh

Maniphest Tasks: T21277

Differential Revision: https://phabricator.tools.flnltd.com/D94732