Fix comment transaction edit policy
Summary: Ref T46181. Tweak the edit policy of comment transaction so that the transaction is editable by the original comment author and //not// the actor that is applying the transactions using the transaction editor.

Test Plan: Migrated a JIRA issue in the staging environment and verified that I was no longer able to edit comments of which I was not the author.

Reviewers: jbeck, tpham

Reviewed By: jbeck

Maniphest Tasks: T46181

Differential Revision: https://phabricator.tools.flnltd.com/D99349