Don't show times for deadline
Summary: The deadline field doesn't currently support times (I wasn't sure if there was a use case so I decided to allow only date-based deadlines for now). However, when we render deadlines we are currently rendering both the date //and// the time (which will always show as "12:00 AM"). Depends on D117358.

Test Plan: Saw times removed from transactions in staging.

Reviewers: jbeck, wjiang, tpham

Reviewed By: jbeck

Differential Revision: https://phabricator.tools.flnltd.com/D117361