Allow searching on the deadline field
Summary: Index the deadline field and allow it to be used with application search. Specifically, allow searching for tasks with deadline before or after a specified time.

Test Plan: Created a task with deadline set to yesterday. Searched for tasks with "Deadline Before" set to `now` and saw expected results.

Reviewers: jbeck, tpham, wjiang

Reviewed By: jbeck

Differential Revision: https://phabricator.tools.flnltd.com/D117354