Prepare deadline custom field for notifications
Summary: I'm working on adding reminder notifications when a deadline is approaching in D117694 but it is becoming quite complicated, especially when considering edge cases, and so I figured I would start splitting some of the changes into smaller diffs. This diff changes the way in which the deadline is stored so that it is JSON encoded, in preparation for storing additional metadata alongside the epoch value.

Test Plan: Ran `bin/storage upgrade` and verified that existing Maniphest tasks with a deadline where still functional. Created a new Maniphest task with a deadline and inspected the `phabricator_maniphest.maniphest_customfieldstorage` table.

Reviewers: jbeck, tpham

Reviewed By: tpham

Differential Revision: https://phabricator.tools.flnltd.com/D118118