Add deadline notifications (mostly)
Summary:
Implements //most// of the code required to generate reminder notifications when a deadline is approaching. Basically what happens here is that whenever the deadline of a Maniphest task is updated we either delete or upsert a trigger. A trigger is an event that is scheduled for some time in the future, which is going to be used to schedule reminder notifications. So that we can modify triggers after they have been created, we need to explicitly keep track of the trigger that we created previously. This is necessary because we can't easily query triggers on action properties, since the `actionProperties` field is JSON-serialized in the database.

Depends on D118118 and D118126.

Test Plan:
# Created a task with no deadline.
  - Checked the database and saw no data stored in `maniphest_customfieldstorage`.
# Changed the deadline to be tomorrow.
  - Checked the database and saw correct data in the `maniphest_customfieldstorage` table.
  - Checked the database and saw correct data in the `worker_trigger` and `worker_triggerevent` tables.
  - Saw "A task deadline is approaching" in the daemon logs.
# Changed the deadline to be today.
 - This did **not** trigger anything, because we are using the one-time trigger clock and it had already fired. This is expected, although I might tweak this behavior in a subsequent diff.
# Removed the task deadline.
  - Checked the database and saw correct data in the `maniphest_customfieldstorage` table.
  - Checked the database and saw that the trigger had been deleted.

Reviewers: jbeck, tpham

Reviewed By: tpham

Differential Revision: https://phabricator.tools.flnltd.com/D118119