Move trigger generation to external transaction effects
Summary: So that `ManiphestDeadlineReminderWorker` can actually comment on a task, I need to pass the `ManiphestTask` PHID to `PhabricatorScheduleTaskTriggerAction`. I can't do this within `applyApplicationTransactionExternalEffects` because the PHID may not have been generated (if creating a new `ManiphestTask`). Instead, pre-generate a PHID for the `PhabricatorWorkerTrigger` in `applyApplicationTransactionInternalEffects` and generate the object itself in `applyApplicationTransactionExternalEffects`.

Test Plan: Created a task in staging and inspected the database.

Reviewers: jbeck, tpham

Reviewed By: jbeck

Differential Revision: https://phabricator.tools.flnltd.com/D118179