Set timeout on RocketChat future
Summary: Ref T76613. Yesterday when dealing with an #incident I attempted to create a ticket and received an execution timeout error message because #phabricator was timing out attempting to trigger #rocket.chat notifications. Add a timeout so that we just give up if we can't post notifications.

Test Plan:
# SSHed to the single `phabricator_web` host in our staging environment and ran `socat TCP-LISTEN:8888,fork SYSTEM:"sleep 60",nofork`.
# Changed `rocketchat.server` to point to `http://localhost:8888`.
# Created a Herald rule to trigger a RocketChat notification on all Maniphest tasks.
# Created a Maniphest task. Before applying this patch, I got a fatal error. After applying this patch, saw no fatal error. Checked the Herald transcript and saw the error message.

{F902303, size=full}

Reviewers: jbeck, tpham

Reviewed By: tpham

Maniphest Tasks: T76613

Differential Revision: https://phabricator.tools.flnltd.com/D118267