Improve deadline reminders
Summary: Ref T76577. I think that the solution that I originally proposed in D118954 doesn't quite work. The issue is that trigger events store the `nextEventEpoch` in the database and is only recalculated when the trigger is fired. As such, modifying a deadline doesn't result in the `nextEventEpoch` field being set to a non-`null` value and, at least in some cases, the trigger doesn't fire. Instead, partially revert D118954 and switch back to using `PhabricatorOneTimeTriggerClock`. When a deadline is modified, delete the trigger event (if one exists) so that the trigger will re-fire. Depends on D118954.

Test Plan:
# Created an event with the deadline set to today.
# Saw deadline reminder.
# Changed deadline to yesterday.
# Saw another deadline reminder.

Reviewers: jbeck, tpham, tmoran

Reviewed By: tpham, tmoran

Maniphest Tasks: T76577

Differential Revision: https://phabricator.tools.flnltd.com/D119878