Filter datasource for Owned projects custom field
Summary: Ref T74624. In this diff we add an alternative datasource for the owned projects customfield in order to apply the constraint that only teams can own other teams or features. This datasource filters out projects that aren't of subtype 'team' or 'feature' with the assumption that the project being created is of subtype 'team'. The new datasource is largely a clone of `PhabricatorProjectDatasource` with a few tweaks. Depends on D126026.

Test Plan:
1. created projects with subtype 'default' and 'team'
2. saw that teams with subtype 'default' were not tokenised

Reviewers: O43 Internal Tooling, O8 Phabricator, josh

Reviewed By: O43 Internal Tooling, O8 Phabricator, josh

Maniphest Tasks: T74624

Differential Revision: https://phabricator.tools.flnltd.com/D125326