[phlab] Check if PHID is empty before querying
Summary:
ref T112443
Currently in staging server, if a project is not owned by any other project, then the found `$owner_project_phids` will be empty and will will break the query later.

This diff checkes the value of `$owner_project_phids`, and only do query if this array is not empty.

Test Plan:
# Previously

1. create a project without owned project
{F1236650, width="70%"}
2. Create a ticket adding this project as tag
{F1236651, width="70%"}
3. When creating, error thrown
{F1236652, width="70%"}

# Now

step 1 and 2 is the same
Ticket is now created successfully
{F1236653, width="70%"}

Reviewers: awright, conrad.olega, O8 Phabricator, tpham

Reviewed By: conrad.olega, O8 Phabricator, tpham

Subscribers: tpham

Maniphest Tasks: T112443

Differential Revision: https://phabricator.tools.flnltd.com/D133371