Add pagerduty custom field
Summary:
ref T109541

Add custom field for specifying the `Pagerduty Service` that will be used for the pagerduty alert. (WIP)

Test Plan:
This is a rather trivial data-like class, and there's no test for custom field throughout the phabricator as well.

I tested this on staging server, and the field can be seen when creating project, and can be read.

**Creating**
{F1260264, width="80%"}

**Displaying**
{F1260267, width="80%"}

**Validation**
{F1271973}

Reviewers: awright, O8 Phabricator, dwang

Reviewed By: awright, O8 Phabricator, dwang

Maniphest Tasks: T109541

Differential Revision: https://phabricator.tools.flnltd.com/D134325