[phlab] fix bug on herald make draft action
Summary:
Ref T233082

The bug has something to do with the number of transactions since we only fetch 20. There's a logic for new revisions not created with `--draft` that require to have two or more `request-review` transactions in order to not be put back to draft. This is because diffs created without flags or with `--only` are immediately published after creating them and the way phab does this internally is with a `request-review` transaction

I have confirmed the behavior by making a diff and making 20+ transactions. Requesting review on that diff will not work due to the logic described above. The fix is to just ignore that logic if the transaction query returns 20 results which is our fetch limit.

Test Plan:
[/] diff then create 20 transactions then request review puts the diff in needs review
[/] new diffs are draft
[/] request review works
[/] updating diffs put it back to draft
[/] unit tests
```
root@b6f2d55f4322:/usr/local/src/phlab# ./../arcanist/bin/arc unit src/applications/herald/action/__tests__/DifferentialMakeDraftHeraldActionTestCase.php
   PASS  165ms   DifferentialMakeDraftHeraldActionTestCase::testIsValidEffect
```

Reviewers: #developer_experience, O8 Phabricator, dwang, jchen

Reviewed By: #developer_experience, O8 Phabricator, dwang, jchen

Maniphest Tasks: T233082

Differential Revision: https://phabricator.tools.flnltd.com/D168674